requirements.txt
absl-py==2.1.0
agora-token-builder==1.0.0
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.6.2
appdirs==1.4.4
apturl==0.5.2
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
billiard==4.2.0
bleach==6.0.0
blinker==1.6.2
blosc==1.9.2
bootstrap4==0.1.0
Bottleneck==1.3.2
Brlapi==0.8.3
Brotli==1.0.9
celery==5.3.6
certifi==2024.2.2
cffi==1.15.1
channels==4.0.0
chardet==4.0.0
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==2.0.0
colorama==0.4.4
command-not-found==0.3
crispy-bootstrap4==2022.1
cron-descriptor==1.4.0
cryptography==3.4.8
cssselect2==0.7.0
cupshelpers==1.0
cvzone==1.6.1
cycler==0.11.0
dask==2022.1.0+dfsg
dataclasses-json==0.6.3
dbus-python==1.2.18
debugpy==1.6.6
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.7
distributed==2022.1.0+ds.1
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.8
django-allauth==0.59.0
django-auto-one-to-one==3.3.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ckeditor==6.7.0
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-js-asset==2.2.0
django-taggit==5.0.1
django-timezone-field==6.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
dropbox==11.26.0
duplicity==0.8.21
et-xmlfile==1.0.1
exceptiongroup==1.2.0
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.3
filelock==3.13.1
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.29.1
fqdn==1.5.1
frozenlist==1.4.1
fs==2.4.12
fsspec==2022.1.0
fusepy==3.0.1
future==0.18.2
gast==0.5.2
greenlet==3.0.3
h11==0.14.0
HeapDict==1.0.1
html5lib==1.1
httplib2==0.20.2
humanfriendly==10.0
idna==3.3
imageio==2.4.1
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==1.1.1
ipykernel==6.17.1
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
isoduration==20.11.0
itsdangerous==2.1.2
jdcal==1.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.1
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.6
keyring==23.5.0
kiwisolver==1.3.2
kombu==5.3.4
langchain==0.0.353
langchain-community==0.0.7
langchain-core==0.1.4
langsmith==0.0.77
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libarchive-c==2.9
llvmlite==0.38.0
locket==0.2.1
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.2
MarkupSafe==2.1.3
marshmallow==3.20.1
mathparse==0.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.6
mediapipe==0.10.9
misaka==2.1.1
mistune==2.0.5
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mpmath==0.0.0
msgpack==1.0.3
multidict==6.0.4
multipledispatch==0.6.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.11
mysqlclient==2.1.1
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==2.4
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numba==0.55.1
numexpr==2.8.1
numpy==1.23.5
oauthlib==3.2.0
odfpy==1.4.2
olefile==0.46
opencv-contrib-python==4.9.0.80
opencv-python==4.6.0.66
openpyxl==3.0.9
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==1.3.5
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
Pint==0.23
platformdirs==4.1.0
pluggy==0.13.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
PyAutoGUI==0.9.54
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.5.3
pydantic_core==2.14.6
PyGetWindow==0.0.9
pygit2==1.6.1
Pygments==2.11.2
PyGObject==3.42.1
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==3.13.0
PyMsgBox==1.0.9
PyMySQL==1.0.3
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==3.14.0
PyPDF2==3.0.1
pyperclip==1.8.2
pypng==0.20220715.0
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.19.3
PyScreeze==0.1.30
PySocks==1.7.1
pytest==6.2.5
python-apt==2.4.0+ubuntu2
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.7.5
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-json-logger==2.0.7
python3-openid==3.2.0
python3-xlib==0.15
pythran==0.10.0
pytweening==1.2.0
pytz==2022.1
PyWavelets==1.1.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.0.2
qrcode==7.4.2
qtconsole==5.4.1
QtPy==2.3.0
regex==2023.6.3
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image==0.18.3
scikit-learn==1.3.0
scipy==1.8.0
SecretStorage==3.3.1
selenium==4.18.1
Send2Trash==1.8.0
six==1.16.0
slugify==0.0.1
sniffio==1.3.0
sortedcontainers==2.1.0
sounddevice==0.4.6
soupsieve==2.3.1
SQLAlchemy==1.2.19
sqlparse==0.4.2
stack-data==0.6.2
stone==3.3.1
svglib==1.5.1
sympy==1.9
systemd-python==234
tables==3.7.0
tblib==1.7.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2022.2.9
tinycss2==1.2.1
toml==0.10.2
toolz==0.11.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.24.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.2.0
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.1.0
virtualenv==20.25.0
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.6
widgetsnbextension==4.0.6
wsproto==1.2.0
xdg==5
xgboost==1.7.6
xhtml2pdf==0.2.11
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
zict==2.0.0
zipp==1.0.0
